requirements.txt
fastapi
uvicorn[standard]
tortoise-orm[asyncpg]
aerich
pydantic[email]
celery[redis]
redis >= 4.0
python-dotenv
requests 
pydantic-settings    
tomlkit
httpx
h2
brotli
dotenv
jmespath
parsel
gevent